+++ packages/article-skeleton/src/article-skeleton.js
@@ -98,7 +98,7 @@ const ArticleSkeleton = ({
       left: 0,
       top: 0
     });
-    window.location.reload();
+    window.location.reload(true);
   };
 
   receiveChildList([
@@ -251,22 +251,21 @@ const ArticleSkeleton = ({
                     isLiveOrBreaking={isLiveOrBreaking}
                   />
                 )}
-                {showUpdate && (
+                {isLiveOrBreaking && showUpdate ? (
                   <TCThemeProvider>
                     <UpdateButtonContainer data-testid="Update button container">
                       <UpdateButtonWithDelay
-                        loading={false}
-                        delay={8000000}
+                        delay={8000}
                         display
-                        label="an update"
+                        label="New update"
                         handleClick={() => scrollToTopAndRefresh(window)}
                         arrowUp
-                        updatedTime={article.updatedTime}
+                        updatedTime={article.publishedTime}
                         articleId={article.id}
                       />
                     </UpdateButtonContainer>
                   </TCThemeProvider>
-                )}
+                ) : null}
                 <PaywallPortal
                   id="paywall-portal-article-footer"
                   componentName="subscribe-cta"