+++ packages/ts-newskit/src/components/misc/delayed-component/__tests__/delayed-component.test.tsx
@@ -5,9 +5,14 @@ import { render, waitFor } from '../../../../utils/test-utils';
 import { DelayedComponent } from '../delayed-component';
 
 describe('DelayedComponent', () => {
+  const mockSetUpdate = jest.fn();
   it('should remove the component after the specified delay when initial state is true', async () => {
     const { queryByText, getByText } = render(
-      <DelayedComponent initialState={true} delay={800}>
+      <DelayedComponent
+        initialState={true}
+        delay={800}
+        setUpdate={mockSetUpdate}
+      >
         <div>Children</div>
       </DelayedComponent>
     );
@@ -18,7 +23,11 @@ describe('DelayedComponent', () => {
   });
   it('should render the component after the specified delay when initial state is false', async () => {
     const { queryByText, getByText } = render(
-      <DelayedComponent initialState={false} delay={800}>
+      <DelayedComponent
+        initialState={false}
+        delay={800}
+        setUpdate={mockSetUpdate}
+      >
         <div>Children</div>
       </DelayedComponent>
     );