+++ packages/ts-newskit/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.65.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.64.0...@times-components/ts-newskit@1.65.0) (2023-07-18)
+
+
+### Features
+
+* **TMRX-1250:** Updates to update button ([#3431](https://github.com/newsuk/times-components/issues/3431)) ([555284d](https://github.com/newsuk/times-components/commit/555284d02bb5e2360843d85cb70e889b969df590))
+
+
+
+
+
+# [1.64.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.63.0...@times-components/ts-newskit@1.64.0) (2023-07-18)
+
+
+### Features
+
+* **TMRX-1250:** Live update button behind flag ([#3430](https://github.com/newsuk/times-components/issues/3430)) ([5069dd6](https://github.com/newsuk/times-components/commit/5069dd62174d9dd63cf78a4b30e8b7ea9582f6b5))
+
+
+
+
+
+# [1.63.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.62.3...@times-components/ts-newskit@1.63.0) (2023-07-17)
+
+
+### Features
+
+* **TMRX-1250:** Live updates behind a flag ([#3429](https://github.com/newsuk/times-components/issues/3429)) ([2fe0200](https://github.com/newsuk/times-components/commit/2fe02005cde5042d7d2ad69c879d0b9a4c0411fc))
+
+
+
+
+
 ## [1.62.3](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.62.2...@times-components/ts-newskit@1.62.3) (2023-07-17)
 
 **Note:** Version bump only for package @times-components/ts-newskit